Create a second version of the password generator
Replacing Random with ThreadLocalRandom allows introducing limit()
method and use of filter(). Due to this, it becomes possible to
generate alpha-numerical password without collecting characters to
Set.

There is still the possibility to generate not a full proper password.
User still is not informed if he / she passed the wrong second startup
parameter.